Added support for psr/log v3 without dropping v2.
Added support for psr/log v3 without dropping v2 by moving setLogger()
implementation to its own package.

* Added dependency on mpdf/psr-log-aware-trait where we created the following:
  * Created trait to handle LoggerAwareInterface setLogger() implementation.
  * Added psr/log v2 support on 2.x branch and v3 support on 3.x branch.
* Widened psr/log version requirement to allow 3.x.
* Updated classes to use MpdfPsrLogAwareTrait in place of direct
  implementation.

on-behalf-of: @ChromaticHQ <mark@chromatichq.com>
