+++ composer.json
@@ -24,11 +24,12 @@
 		"php": "^5.6 || ^7.0 || ~8.0.0 || ~8.1.0 || ~8.2.0",
 		"ext-gd": "*",
 		"ext-mbstring": "*",
+		"mpdf/psr-log-aware-trait": "^2.0 || ^3.0",
 		"myclabs/deep-copy": "^1.7",
 		"paragonie/random_compat": "^1.4|^2.0|^9.99.99",
 		"php-http/message-factory": "^1.0",
 		"psr/http-message": "^1.0",
-		"psr/log": "^1.0 || ^2.0",
+		"psr/log": "^1.0 || ^2.0 || ^3.0",
 		"setasign/fpdi": "^2.1"
 	},
 	"require-dev": {
