+++ phpunit.xml
@@ -6,7 +6,7 @@
 
 	<testsuites>
 		<testsuite name="Tests">
-			<directory suffix=".php">./tests</directory>
+			<directory suffix="Test.php">./tests</directory>
 		</testsuite>
 	</testsuites>
