+++ src/Http/SocketHttpClient.php
@@ -3,13 +3,14 @@
 namespace Mpdf\Http;
 
 use Mpdf\Log\Context as LogContext;
+use Mpdf\PsrLogAwareTrait\PsrLogAwareTrait;
 use Psr\Http\Message\RequestInterface;
 use Psr\Log\LoggerInterface;
 
 class SocketHttpClient implements \Mpdf\Http\ClientInterface, \Psr\Log\LoggerAwareInterface
 {
 
-	private $logger;
+	use PsrLogAwareTrait;
 
 	public function __construct(LoggerInterface $logger)
 	{
@@ -102,11 +103,6 @@ class SocketHttpClient implements \Mpdf\Http\ClientInterface, \Psr\Log\LoggerAwa
 
 		return $response
 			->withBody($stream);
-	}
-
-	public function setLogger(LoggerInterface $logger)
-	{
-		$this->logger = $logger;
 	}
 
 }
