+++ src/Image/ImageProcessor.php
@@ -13,11 +13,14 @@ use Mpdf\Language\ScriptToLanguageInterface;
 use Mpdf\Log\Context as LogContext;
 use Mpdf\Mpdf;
 use Mpdf\Otl;
+use Mpdf\PsrLogAwareTrait\PsrLogAwareTrait;
 use Mpdf\SizeConverter;
 use Psr\Log\LoggerInterface;
 
 class ImageProcessor implements \Psr\Log\LoggerAwareInterface
 {
+
+	use PsrLogAwareTrait;
 
 	/**
 	 * @var \Mpdf\Mpdf
@@ -89,11 +92,6 @@ class ImageProcessor implements \Psr\Log\LoggerAwareInterface
 	 */
 	private $assetFetcher;
 
-	/**
-	 * @var \Psr\Log\LoggerInterface
-	 */
-	public $logger;
-
 	public function __construct(
 		Mpdf $mpdf,
 		Otl $otl,
@@ -124,18 +122,6 @@ class ImageProcessor implements \Psr\Log\LoggerAwareInterface
 		$this->guesser = new ImageTypeGuesser();
 
 		$this->failedImages = [];
-	}
-
-	/**
-	 * @param \Psr\Log\LoggerInterface
-	 *
-	 * @return self
-	 */
-	public function setLogger(LoggerInterface $logger)
-	{
-		$this->logger = $logger;
-
-		return $this;
 	}
 
 	public function getImage(&$file, $firstTime = true, $allowvector = true, $orig_srcpath = false, $interpolation = false)
