+++ src/Writer/MetadataWriter.php
@@ -6,6 +6,7 @@ use Mpdf\Strict;
 use Mpdf\Form;
 use Mpdf\Mpdf;
 use Mpdf\Pdf\Protection;
+use Mpdf\PsrLogAwareTrait\PsrLogAwareTrait;
 use Mpdf\Utils\PdfDate;
 
 use Psr\Log\LoggerInterface;
@@ -14,6 +15,7 @@ class MetadataWriter implements \Psr\Log\LoggerAwareInterface
 {
 
 	use Strict;
+	use PsrLogAwareTrait;
 
 	/**
 	 * @var \Mpdf\Mpdf
@@ -34,11 +36,6 @@ class MetadataWriter implements \Psr\Log\LoggerAwareInterface
 	 * @var \Mpdf\Pdf\Protection
 	 */
 	private $protection;
-
-	/**
-	 * @var \Psr\Log\LoggerInterface
-	 */
-	private $logger;
 
 	public function __construct(Mpdf $mpdf, BaseWriter $writer, Form $form, Protection $protection, LoggerInterface $logger)
 	{
@@ -806,11 +803,6 @@ class MetadataWriter implements \Psr\Log\LoggerAwareInterface
 			$uniqid = md5(time() . $this->mpdf->buffer);
 			$this->writer->write('/ID [<' . $uniqid . '> <' . $uniqid . '>]');
 		}
-	}
-
-	public function setLogger(LoggerInterface $logger)
-	{
-		$this->logger = $logger;
 	}
 
 	private function getVersionString()
